Manejo de estado
Se completa la implementación de métodos en HomeBloc.
Implementación de BlocBuilder en HomeScreen.
Se refactoriza las funcionalidades de Repository en CitiesRepository y GeographicDataService
