+++ lib/components/homeScreen/bloc/home_bloc.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
+import 'package:goodmeal_test/core/repositories/citiesRepository.dart';
 import 'package:meta/meta.dart';
 
 part 'home_event.dart';
@@ -8,6 +9,8 @@ part 'home_state.dart';
 
 class HomeBloc extends Bloc<HomeEvent, HomeState> {
   HomeBloc() : super(HomeUninitialized());
+
+  CitiesRepository _citiesRepository = CitiesRepository.instance;
 
   @override
   Stream<HomeState> mapEventToState(
@@ -18,8 +21,9 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
     }
   }
 
-
-  Stream<HomeState> _mapLoadDataToState() async *{
-    //TODO: Implement Method
+  Stream<HomeState> _mapLoadDataToState() async* {
+    yield HomeLoading();
+    await _citiesRepository.loadData();
+    yield HomeInitialized();
   }
 }
