+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:goodmeal_test/components/homeScreen/view/homeView.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:goodmeal_test/components/homeScreen/bloc/home_bloc.dart';
+import 'package:goodmeal_test/components/homeScreen/view/homeScreen.dart';
 
 void main() {
   runApp(WeatherApp());
@@ -15,54 +17,10 @@ class WeatherApp extends StatelessWidget {
         fontFamily: 'Montserrat',
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: HomeScreen(),
-    );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  HomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headline4,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ),
+      home: BlocProvider(
+        create: (context) => HomeBloc()..add(LoadData()),
+        lazy: false,
+        child: HomeScreen()),
     );
   }
 }
